langchain_summarizer/pyproject.toml
[tool.poetry]
name = "langchain-summarizer"
version = "0.1.0"
description = ""
authors = ["Vikram <vikramsingh001@gmail.com>"]
readme = "README.md"
packages = [{include = "langchain_summarizer"}]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0" 
langchain = "^0.0.188"
transformers = "^4.29.2"
python-dotenv = "^1.0.0"
text-generation = "^0.5.2"
openai = "^0.27.8"
tiktoken = "^0.4.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
types-requests = "^2.31.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
